requirements.txt
mecab-python3 == 0.996.2
tqdm == 4.32.1
